refactor: organize desktop app source into app and core directories
