+++ apps/desktop/src/app/App.tsx
@@ -21,29 +21,29 @@ import { getCurrentWindow } from "@tauri-apps/api/window";
 import { open } from "@tauri-apps/plugin-dialog";
 import { openUrl } from "@tauri-apps/plugin-opener";
 import { Activity, Award, Bot, Boxes, Brain, Check, ChevronRight, CircleAlert, Code2, Copy, ExternalLink, FileCode2, FolderGit2, Gauge, GitCommitHorizontal, GitCompareArrows, History, Home, LayoutDashboard, Library, MessageSquareText, MoreHorizontal, Pencil, PlugZap, RefreshCw, Search, ShieldCheck, Sparkles, Trash2, X } from "lucide-react";
-import { api } from "./api";
+import { api } from "../core/api";
 import { cn } from "@/lib/utils";
-import { AgentIcon } from "./components/AgentIcon";
-import { AppSidebar, type SidebarEntry } from "./components/AppSidebar";
-import { SettingsSidebar, type SettingsSection } from "./components/SettingsSidebar";
-import { WindowToolbar } from "./components/WindowToolbar";
-import { ObsidianSettingsCard, WorkspaceObsidianCard } from "./components/ObsidianIntegration";
-import { RemoteGatewaysSettings } from "./components/RemoteGateways";
-import { QuotaDiagnostics } from "./components/QuotaDiagnostics";
-import { WorkspaceStoragePage } from "./components/WorkspaceStoragePage";
-import { WorkspaceSessionsPage } from "./components/WorkspaceSessionsPage";
-import { WorkspaceDoctorPage } from "./components/WorkspaceDoctorPage";
-import { WorkspaceGitPage, type GitSubview } from "./components/WorkspaceGitPage";
-import { WorkspaceOpenWith } from "./components/WorkspaceOpenWith";
-import { AssetCatalogPage } from "./components/AssetCatalogPage";
-import { useAppDialogs } from "./components/AppDialogProvider";
-import type { InsightsSection } from "./components/InsightsPage";
-import { groupCatalogAssets, groupWorkspaceAssets, workspaceAssetCounts, type CatalogAssetGroup } from "./catalog";
-import { diffLines } from "./diff";
-import { changeLocale, formatCompactNumber, formatDateTime, formatRelativeTime, localizeMessage, tr } from "./i18n";
-import { applyTheme } from "./theme";
-import { normalizePlatform, primaryShortcutModifier, usesSystemTrayWording } from "./platform";
-import type { ActivityRecord, AgentInstallation, AgentKind, AppIconPreference, AppMenuCommandRequest, AppNavigationRequest, CatalogAsset, ChangeSet, CloseBehavior, ConnectionDefinition, ContextDoctorSummary, ContextPreview, DiscoveryReport, EffectiveTheme, ExcludedWorkspace, GitIdentitySummary, InsightsStatus, InsightsSummary, LocalePreference, Manifest, McpInstallation, McpRegistryEntry, McpRuntimeStatus, McpServerConfig, MemoryRecord, MemoryType, QuotaCollectorStatus, QuotaWindowSelector, RefreshJobStatus, RefreshKind, RemoteGatewaySummary, RuntimeInfo, ScanRoot, SessionHandoffLaunchRequest, ThemePreference, WorkspaceScan, WorkspaceSummary } from "./types";
+import { AgentIcon } from "../components/AgentIcon";
+import { AppSidebar, type SidebarEntry } from "../components/AppSidebar";
+import { SettingsSidebar, type SettingsSection } from "../components/SettingsSidebar";
+import { WindowToolbar } from "../components/WindowToolbar";
+import { ObsidianSettingsCard, WorkspaceObsidianCard } from "../components/ObsidianIntegration";
+import { RemoteGatewaysSettings } from "../components/RemoteGateways";
+import { QuotaDiagnostics } from "../components/QuotaDiagnostics";
+import { WorkspaceStoragePage } from "../components/WorkspaceStoragePage";
+import { WorkspaceSessionsPage } from "../components/WorkspaceSessionsPage";
+import { WorkspaceDoctorPage } from "../components/WorkspaceDoctorPage";
+import { WorkspaceGitPage, type GitSubview } from "../components/WorkspaceGitPage";
+import { WorkspaceOpenWith } from "../components/WorkspaceOpenWith";
+import { AssetCatalogPage } from "../components/AssetCatalogPage";
+import { useAppDialogs } from "../components/AppDialogProvider";
+import type { InsightsSection } from "../components/InsightsPage";
+import { groupCatalogAssets, groupWorkspaceAssets, workspaceAssetCounts, type CatalogAssetGroup } from "../core/catalog";
+import { diffLines } from "../core/diff";
+import { changeLocale, formatCompactNumber, formatDateTime, formatRelativeTime, localizeMessage, tr } from "../core/i18n";
+import { applyTheme } from "../core/theme";
+import { normalizePlatform, primaryShortcutModifier, usesSystemTrayWording } from "../core/platform";
+import type { ActivityRecord, AgentInstallation, AgentKind, AppIconPreference, AppMenuCommandRequest, AppNavigationRequest, CatalogAsset, ChangeSet, CloseBehavior, ConnectionDefinition, ContextDoctorSummary, ContextPreview, DiscoveryReport, EffectiveTheme, ExcludedWorkspace, GitIdentitySummary, InsightsStatus, InsightsSummary, LocalePreference, Manifest, McpInstallation, McpRegistryEntry, McpRuntimeStatus, McpServerConfig, MemoryRecord, MemoryType, QuotaCollectorStatus, QuotaWindowSelector, RefreshJobStatus, RefreshKind, RemoteGatewaySummary, RuntimeInfo, ScanRoot, SessionHandoffLaunchRequest, ThemePreference, WorkspaceScan, WorkspaceSummary } from "../core/types";
 
 type Page = "overview" | "sessions" | "git" | "assets" | "context" | "doctor" | "changes";
 type GlobalPage = "home" | "workspaces" | "catalog" | "agents" | "quota" | "insights";
@@ -53,9 +53,9 @@ type WorkspaceAssetSection = "instructions" | "skills" | "mcp" | "native";
 type WorkspaceView = "list" | "storage";
 type ChangeSetOrigin = "standard" | "doctor" | "handoff";
 
-const AgentsPageLazy = lazy(() => import("./components/AgentsPage").then(({ AgentsPage }) => ({ default: AgentsPage })));
-const InsightsPageLazy = lazy(() => import("./components/InsightsPage").then(({ InsightsPage }) => ({ default: InsightsPage })));
-const QuotaPageLazy = lazy(() => import("./components/QuotaPage").then(({ QuotaPage }) => ({ default: QuotaPage })));
+const AgentsPageLazy = lazy(() => import("../components/AgentsPage").then(({ AgentsPage }) => ({ default: AgentsPage })));
+const InsightsPageLazy = lazy(() => import("../components/InsightsPage").then(({ InsightsPage }) => ({ default: InsightsPage })));
+const QuotaPageLazy = lazy(() => import("../components/QuotaPage").then(({ QuotaPage }) => ({ default: QuotaPage })));
 const buildPlatform = import.meta.env.TAURI_ENV_PLATFORM;
 const appPlatform = normalizePlatform(buildPlatform);
 const hasFileAccessSettings = ["macos", "windows"].includes(appPlatform);
@@ -678,7 +678,7 @@ function McpServerEditor({ project, onSaved }: { project?: string; onSaved: () =
 
 function McpMigrationInventory({ project, onPlanned }: { project?: string; onPlanned: (project: string, changeSet: ChangeSet) => Promise<void> }) {
   const dialogs = useAppDialogs();
-  const [candidates, setCandidates] = useState<import("./types").McpMigrationCandidate[]>([]);
+  const [candidates, setCandidates] = useState<import("../core/types").McpMigrationCandidate[]>([]);
   const [scanned, setScanned] = useState(false);
   const [selected, setSelected] = useState<string[]>([]);
   const [busy, setBusy] = useState(false);
