+++ apps/desktop/src/components/AgentIcon.test.tsx
@@ -2,7 +2,7 @@
 import { cleanup, render } from "@testing-library/react";
 import { afterEach, describe, expect, it } from "vitest";
 import { AgentIcon } from "./AgentIcon";
-import type { AgentKind } from "../types";
+import type { AgentKind } from "../core/types";
 
 afterEach(cleanup);
