+++ apps/desktop/src/components/AppDialogProvider.test.tsx
@@ -3,7 +3,7 @@ import "@testing-library/jest-dom/vitest";
 import { cleanup, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
-import { initializeI18n } from "../i18n";
+import { initializeI18n } from "../core/i18n";
 import { Button } from "./ui/button";
 import { AppDialogProvider, useAppDialogs } from "./AppDialogProvider";
