+++ apps/desktop/src/components/AppDialogProvider.tsx
@@ -14,7 +14,7 @@ import { Button } from "@/components/ui/button";
 import { Dialog, DialogContent, DialogDescription, DialogFooter, DialogHeader, DialogTitle } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { tr } from "../i18n";
+import { tr } from "../core/i18n";
 
 type DialogTone = "default" | "destructive" | "warning";
