+++ apps/desktop/src/components/QuotaPage.test.tsx
@@ -2,7 +2,7 @@
 import "@testing-library/jest-dom/vitest";
 import { cleanup, fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { initializeI18n } from "../i18n";
+import { initializeI18n } from "../core/i18n";
 import { QuotaPage } from "./QuotaPage";
 
 const { listen, quotaSnapshot, quotaCollectorStatus, quotaPopoverPreferences, setQuotaPopoverPreferences, refreshQuota, refreshStatus, requestRefresh } = vi.hoisted(() => ({
@@ -16,7 +16,7 @@ const { listen, quotaSnapshot, quotaCollectorStatus, quotaPopoverPreferences, se
   requestRefresh: vi.fn(),
 }));
 vi.mock("@tauri-apps/api/event", () => ({ listen }));
-vi.mock("../api", () => ({
+vi.mock("../core/api", () => ({
   api: { quotaSnapshot, quotaCollectorStatus, quotaPopoverPreferences, setQuotaPopoverPreferences, refreshQuota, refreshStatus, requestRefresh },
 }));
