+++ apps/desktop/src/core/achievements.test.ts
