+++ apps/desktop/src/core/achievements.ts
