+++ apps/desktop/src/core/catalog.test.ts
