+++ apps/desktop/src/core/diff.ts
