+++ apps/desktop/src/core/i18n.ts
@@ -1,9 +1,9 @@
 import i18n, { type TOptions } from "i18next";
 import { initReactI18next } from "react-i18next";
-import enUS from "./locales/en-US.json";
-import jaJP from "./locales/ja-JP.json";
-import zhCN from "./locales/zh-CN.json";
-import zhTW from "./locales/zh-TW.json";
+import enUS from "../locales/en-US.json";
+import jaJP from "../locales/ja-JP.json";
+import zhCN from "../locales/zh-CN.json";
+import zhTW from "../locales/zh-TW.json";
 import type { LocalizedMessage, SupportedLocale } from "./types";
 
 export type LocalePreference = "system" | SupportedLocale;
