+++ apps/desktop/src/core/insights.test.ts
