+++ apps/desktop/src/core/insights.ts
