+++ apps/desktop/src/core/platform.test.ts
