+++ apps/desktop/src/core/platform.ts
