+++ apps/desktop/src/core/types.ts
