+++ apps/desktop/src/test/ui-source-constraints.test.ts
@@ -3,7 +3,7 @@ import { dirname, join } from "node:path";
 import { fileURLToPath } from "node:url";
 import { describe, expect, it } from "vitest";
 
-const sourceRoot = dirname(fileURLToPath(import.meta.url));
+const sourceRoot = join(dirname(fileURLToPath(import.meta.url)), "..");
 
 function businessSources(directory: string): string[] {
   return readdirSync(directory, { withFileTypes: true }).flatMap((entry) => {
